CheckIfTypeIsEquivalent POC
Added a secondary cache in EquivalentCache that holds the 8 most recently seen non-shared types for which we have returned that the  type is non-equivalent. If we see this type again within next 8 tries, we will cache it.

Also fixed bug around nonSharedSlotIndex. Today we always pick the 1st non-shared type we see in equivTypes[]. With that we will be always evicting recently cached non-shared type.